Add endpoint to fetch assignments for a specific student
A new HTTP GET endpoint has been added to the `AssignmentsController`
class to retrieve assignments for a specific student. The endpoint
is defined with the route `student/{student_id}` and is handled by
the `GetAssignmentsForStudent` method. This method constructs a SQL
query to select assignments associated with the specified student
and ensures they are not marked as deleted. The query results are
mapped to a list of `Assignments` objects and returned as an HTTP
200 OK response.